style: Rename `Primitives` to `IngressPrimitives` and `EPrimitives` to `EgressPrimitives`
Ingress means "entering the router" and Egress means "leaving the router".

This change could of course be reverted (through r-a) right before merging
the refactor/liveliness-declaration branch.

It would've been even better to move Session's IngressPrimitives impl to an
EgressPrimitives, but that's not possible since a Session can recieve Query
replies from itself.

Ideally, Session in/egress should be Router e/ingress but the `EPrimitives`
trait breaks this symmetry. Messages leaving the router are not the same as
messages leaving the Session. This could be solved by making the `EgressPrimitives`
trait generic over the message context, but the necessary refactoring would
be a hude undertaking, and I don't have enough time nor familiarity with the
codebase for it.