fix(replication): do not reply an error when a key is not found
As discussed internally, a reply to a query should not return an error
when a key does not exist — this is a perfectly valid situation.

In addition, this prevented the replication from correctly functioning:
as an error was always returned by the storage that was actually
looking for the data, other replies were not processed and that storage
would thus never receive anything.

* plugins/zenoh-plugin-storage-manager/src/replica/storage.rs: deleted
  the code replying an error to a query looking for a key the storage
  does not have.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>