feat!: make callback an opaque type
It may allow special case for channel handlers later, in order to use async
methods, or change the return type of the callback in order to support
async callbacks.

For the record I've tested quickly with an enum inside `Callback`,
and there was a sensitive cost in performance of 1% when using callbacks.