refactor(storage-manager): keep configuration in StorageService
The `StorageService` was splitting the `StorageConfig` that was used to
create it. In addition to adding noise, this prevented separating the
creation of the structure from spawning the subscriber and queryable
associated with a Storage.

This commit changes the fields of the `StorageService` structure to keep
the entire `StorageConfig` -- thus allowing separating the creation from
spawning the queryable and subscriber.

* plugins/zenoh-plugin-storage-manager/src/storages_mgt/service.rs:
  - Removed the following fields from the `StorageService` structure:
    - `key_expr`,
    - `complete`,
    - `strip_prefix`.
  - Added the field `configuration` to keep track of the associated
    `StorageConfig`.
  - Changed the signature of the `start_storage_queryable_subscriber`
    removing the `GarbageConfig` as it is now contained in `&self`.
  - Updated the calls to access `key_expr`, `complete` and
    `strip_prefix`.
  - Removed an `unwrap()` and instead log an error.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>