refactor(storage-manager): `reply_events` -> `reply_event_retrieval`
This commit refactors, renames and changes the method `reply_events` to
`reply_event_retrieval`: it now only sends a single
`AlignmentReply::Retrieval` to the Replica that requested data.

The code of the method was also reworked to make it clearer and less
error-prone: there are no `continue;` statement in the middle which
makes the flow less "jumpy".

These changes prepare the introduction of the Wildcard Update.

* plugins/zenoh-plugin-storage-manager/src/replication/core/aligner_query.rs:
  refactor, rename and change the method `reply_events` to
  `reply_event_retrieval`.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>