fix: don't emit an error log if pipeline is closed
Before, there was no distinction between the pipeline dropping a
message because its deadline was reached or because the transport was
closed.
As a consequence, blocking message which was dropped because of
closed transport would still emit an error log. This PR introduce
the distinction between the two dropping causes.