Implement pub/sub, query/reply and liveliness gateway routing
At this stage, the interest protocol is unimplemented. All hats receive
entities known by the gateway, but currently do not respect PULL mode in
south-bound subregions. Interest routing methods are stubs.

Isolation of south and north linkstate regions is handled by forwarding
sourced entities from other regions using the gateway as a source.

There is no mechanism yet to elect primary gateways in case a region
has multiple gateways, so data may be duplicated.

Sessions are north-bound as the router and peer hats retain the ability
to route for clients.

There is still no way to configure the south hat(s) and thus gateways
remain untestable.

An additional "eastwest" bound is devised to handle nodes that belong
neither to the north region nor the south subregions. Such nodes are
handled as if they were south-bound in practice.

The adminspace is unimplemented for now.

Failover brokering is unimplemented for now.