Implement interest gateway routing
Router networks currently only support one gateway. Peer networks
consider all routers to be gateways, but it remains to put in a
place mechanisms to (1) advertise south-bound peers as gateways
and (2) perform elections to decide which peer gateway would
propagate declarations dowstream. (2) will likely not be implemented
in the near future.