refactor RoutingExpr to avoid useless allocation
When there is no suffix, or when a resource exists for the full expr,
there is no point in reallocating the full string as it can be obtained
from the resource.
The computation of the full expr resource is also cached, as it was done
multiple times.

A malicious bug has also been fixed in `route_data`, as the prefix was
cloned for no relevant reason, so it was still alive when the table lock
was released.