Close the advanced subscriber with the session
It also fixes the bug of the session not being closed when an advanced
subscriber is alive.

WeakSession in the code becomes a simple alias to Arc<SessionInner>
thanks to #1641. A new WeakSessionInternal, exported as WeakSession, can
be used in internal crates, such as zenoh-ext, to keep a weak session
reference. So the advanced subscriber only captures a weak session now.

Moreover, in order to properly close the advanced subscriber, its
callback must be dropped. It is done in the drop callback of the
underlying subscriber, so the advanced subscriber is now really bound
to the real registered subscriber.