Panic when sample_depth is zero in advanced pub/sub
A history depth of zero makes no sense, both on publisher cache side,
and on subscriber query side.

There was a bug in the subscriber, as not having `sample_depth` set
was preventing proper buffering. It has been fixed by considering that
no `sample_depth` set means an arbitrary number of responses to the
query, so responses must be buffered till the end.